pyproject.toml
[tool.poetry]
name = "spa-web-applications"
version = "0.1.0"
description = ""
authors = ["Aristov Sergei <maxmit83@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
ipython = "^8.25.0"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
celery = "^5.4.0"
python-telegram-bot = "^21.3"
redis = "^5.0.7"
django-celery-beat = "^2.6.0"
drf-yasg = "^1.21.7"
setuptools = "^70.2.0"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
